댓글 기능 및 css 적용

+++ articles/views.py
@@ -1,10 +1,11 @@
-from django.shortcuts import render, redirect
+from gc import get_objects
+from django.shortcuts import render, redirect, get_object_or_404
 from django.views.decorators.http import require_http_methods, require_POST, require_safe
 from django.contrib.auth.decorators import login_required
 from django.views.generic.edit import FormView
 from django.urls import reverse
-from .models import Article
-from .forms import ArticleForm
+from .models import Article, Comment
+from .forms import ArticleForm, CommentForm
 
 # Create your views here.
 @require_safe
@@ -21,7 +22,9 @@ def create(request):
     if request.method == 'POST':
         form = ArticleForm(request.POST)
         if form.is_valid():
-            article = form.save()
+            article = form.save(commit=False)
+            article.user = request.user
+            article.save()
             return redirect('articles:detail', article.pk)
     else:
         form = ArticleForm()
@@ -34,18 +37,21 @@ def create(request):
 @require_safe
 def detail(request, pk):
     article = Article.objects.get(pk=pk)
+    comment_form = CommentForm()
+    comments = Comment.objects.all()
     context = {
         'article': article,
+        'comments': comments,
+        'comment_form': comment_form,
     }
     return render(request, 'articles/detail.html', context)
 
 @login_required
 @require_http_methods(['GET', 'POST'])
 def update(request, pk):
-    article = Article.objects.get(pk=pk)
+    article = get_object_or_404(Article, pk=pk)
     if request.method == 'POST':
         form = ArticleForm(request.POST, instance=article)
-        # form = ArticleForm(data=request.POST, instance=article)
         if form.is_valid():
             form.save()
             return redirect('articles:detail', article.pk)
@@ -60,6 +66,32 @@ def update(request, pk):
 @login_required
 @require_POST
 def delete(request, pk):
-    article = Article.objects.get(pk=pk)
-    article.delete()
-    return redirect('articles:index')
+    article = get_object_or_404(Article, pk=pk)
+    if request.user.is_authenticated:
+        if request.user == article.user:
+            article.delete()
+            return redirect('articles:index')
+    return redirect('articles:detail', article.pk)
+
+
+# 여기서부터 댓글 부분
+@require_POST
+def comments_create(request, pk):
+    if request.user.is_authenticated:
+        article = get_object_or_404(Article, pk=pk)
+        comment_form = CommentForm(request.POST)
+        if comment_form.is_valid():
+            comment = comment_form.save(commit=False)
+            comment.article = article
+            comment.user = request.user
+            comment.save()
+        return redirect('articles:detail', article.pk)
+    return redirect('accounts:login')
+
+@require_POST
+def comments_delete(request, article_pk, comment_pk):
+    if request.user.is_authenticated:
+        comment = get_object_or_404(Comment, pk=comment_pk)
+        if request.user == comment.user:
+            comment.delete()
+    return redirect('article:detail', article_pk)
